✅ 리뷰 좋아요 상태 확인 API 핸들러 통합 및 컴포넌트 구조 개선
- 기존의 post.handler.ts 파일을 route.ts로 통합하여 API 핸들러 구조를 단순화하고 코드 가독성을 향상시킴

- 최신 리뷰 목록을 표시하는 LatestReviewList 컴포넌트로 이름 변경

- 리뷰 티켓 컴포넌트에서 불필요한 주석 및 코드 정리를 통해 가독성 개선

- 파일 업로드 관련 유틸리티 함수 및 상수를 별도의 파일로 분리하여 코드 구조 개선
