+++ app/components/ui/buttons/WriteBtn.tsx
@@ -12,7 +12,7 @@ export default function WriteBtn({ movieId, size = "large" }: WriteBtnProps) {
       href={`/write-review/new?movieId=${movieId}`}
       className={`${
         size === "large"
-          ? "rounded-2xl p-3 text-sm md:text-base"
+          ? "rounded-2xl p-4 text-sm md:text-base"
           : "rounded-xl px-3 py-2 text-xs"
       } group relative flex w-full items-center justify-between overflow-hidden bg-gradient-to-br from-primary-400 via-primary-600 to-primary-800 text-white transition-all duration-300 hover:from-primary-500 hover:via-primary-700 hover:to-primary-900 hover:shadow-lg focus:outline-none focus:ring-1 focus:ring-accent-300 focus:ring-offset-1`}
     >
