+++ app/home/components/index.ts
@@ -1,4 +1,4 @@
 export { default as RecommendSection } from "app/home/components/RecommendSection";
 export { default as RecommendMovie } from "app/home/components/RecommendMovie";
 export { default as MovieSection } from "app/home/components/MovieSection";
-export { default as LatestReviews } from "app/home/components/reviews/LatestReviews";
+export { default as LatestReviewList } from "app/home/components/reviews/LatestReviewList";
