+++ app/home/home-page.tsx
@@ -2,15 +2,15 @@
 
 import dynamic from "next/dynamic";
 import { RecommendSection, MovieSection } from "app/home/components";
-import LatestReviewsSkeleton from "app/home/components/reviews/LatestReviewsSkeleton";
+import LatestReviewSkeleton from "app/home/components/reviews/LatestReviewSkeleton";
 import { MovieList } from "lib/movies/fetchNowPlayingMovies";
 import { ReviewDoc } from "lib/reviews/fetchReviewsPaginated";
 
 // 컴포넌트 지연 로딩
-const LatestReviews = dynamic(
-  () => import("app/home/components/reviews/LatestReviews"),
+const LatestReviewList = dynamic(
+  () => import("app/home/components/reviews/LatestReviewList"),
   {
-    loading: () => <LatestReviewsSkeleton />,
+    loading: () => <LatestReviewSkeleton />,
   },
 );
 
@@ -30,22 +30,20 @@ export default function HomePage({
   latestReviews,
 }: HomePageProps) {
   return (
-    <main className="relative">
-      <div className="px-4 md:px-6 lg:px-8">
-        <MovieSection
-          title="Now Playing"
-          description="지금 상영 중인 영화들을 확인하고 티켓을 예매하세요"
-          movieList={movieList}
-          maxItems={10}
-        />
-        <RecommendSection movie={recommendMovie} trailerKey={trailerKey} />
-        <MovieSection
-          title="Trending Movies"
-          description="요즘 가장 인기 있는 영화들을 만나보세요"
-          movieList={trendingMovies}
-        />
-        <LatestReviews reviews={latestReviews} />
-      </div>
+    <main className="px-4 md:px-6 lg:px-8">
+      <MovieSection
+        title="Now Playing"
+        description="지금 상영 중인 영화들을 확인하고 티켓을 예매하세요"
+        movieList={movieList}
+        maxItems={10}
+      />
+      <RecommendSection movie={recommendMovie} trailerKey={trailerKey} />
+      <MovieSection
+        title="Trending Movies"
+        description="요즘 가장 인기 있는 영화들을 만나보세요"
+        movieList={trendingMovies}
+      />
+      <LatestReviewList reviews={latestReviews} />
     </main>
   );
 }
