+++ app/movie-details/[id]/components/MetaInfoItem.tsx
@@ -7,7 +7,7 @@ interface MetaInfoItemProps {
 
 const MetaInfoItem = ({ label, children }: MetaInfoItemProps) => (
   <div className="flex flex-1 flex-col">
-    <h3 className="mb-2 text-xs font-bold">{label}</h3>
+    <h3 className="mb-1 text-xs font-bold">{label}</h3>
     <div className="break-keep text-center text-sm">{children}</div>
   </div>
 );
