+++ app/movie-details/[id]/page.tsx
@@ -1,14 +1,14 @@
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
-import MovieDetailCard from "app/movie-details/components/MovieDetails";
-import AllMovieTrailers from "app/movie-details/components/MovieTrailers";
-import SimilarMovies from "app/movie-details/components/SimilarMovies";
+import MovieDetailCard from "app/movie-details/[id]/components/MovieDetails";
+import MovieTrailerList from "app/movie-details/[id]/components/MovieTrailerList";
 import Background from "app/components/ui/layout/Background";
 import getMovieTitle from "app/utils/getMovieTitle";
 import { fetchMovieCredits } from "lib/movies/fetchMovieCredits";
 import { fetchMovieDetails } from "lib/movies/fetchMovieDetails";
 import { fetchSimilarMovies } from "lib/movies/fetchSimilarMovies";
 import { fetchVideosMovies } from "lib/movies/fetchVideosMovies";
+import SimilarMovieList from "app/movie-details/[id]/components/SimilarMovieList";
 
 export async function generateMetadata({
   params,
@@ -82,8 +82,8 @@ export default async function MovieDetailPage({
           movieDetails={movieDetails}
           movieCredits={movieCredits}
         />
-        <AllMovieTrailers movieTrailer={movieTrailerData.results} />
-        <SimilarMovies similarMovies={similarMovies} />
+        <MovieTrailerList trailerList={movieTrailerData.results} />
+        <SimilarMovieList movieList={similarMovies} />
       </>
     );
   } catch (error: unknown) {
