+++ app/my-page/components/profile-avatar/utils/allowedExtensions.ts
@@ -1,0 +1,2 @@
+// 허용된 파일 확장자 목록
+export const ALLOWED_EXTENSIONS = [".jpg", ".jpeg", ".png", ".gif"];
