+++ app/my-page/components/profile-avatar/utils/formatFileSize.ts
@@ -1,0 +1,8 @@
+// 파일 크기를 읽기 쉬운 형태로 변환
+export const formatFileSize = (bytes: number): string => {
+  if (bytes === 0) return "0 Bytes";
+  const k = 1024;
+  const sizes = ["Bytes", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + " " + sizes[i];
+};
